+++ src/main/java/view/PopUpButton.java
@@ -40,7 +40,7 @@ public class PopUpButton extends Drawable implements Clickable  {
         private float opacity = 0.0f;
         @Override
         public boolean isIn(MouseEvent e) {
-            return Math.sqrt(Math.pow(e.getX() - x, 2) + Math.pow((e.getY() - y), 2)) < r;
+            return Math.sqrt(Math.pow((double)e.getX() - x, 2) + Math.pow(((double) e.getY() - y), 2)) < r;
         }
 
         @Override
