+++ src/main/java/view/SpringView.java
@@ -101,7 +101,7 @@ public class SpringView extends Drawable implements Clickable, CreatePopUpBar, S
      */
     @Override
     public boolean isIn(MouseEvent e) {
-        return Math.sqrt(Math.pow(e.getX() - x, 2) + Math.pow((e.getY() - y), 2)) < r;
+        return Math.sqrt(Math.pow((double) e.getX() - x, 2) + Math.pow(((double) e.getY() - y), 2)) < r;
     }
     @Override
     public void clickAction(MouseEvent e) {
